+++ scripts/lib/socket.js
@@ -1,12 +1,12 @@
 import { MODULE_ID } from "../const.js";
-import { handleSummonPlacingAndPost } from "../module.js";
+import { handleSummonPlacementAndPost } from "../summon.js";
 
 let socketlibSocket;
 // Taken by Reference from Pf2e Action Support
 export const setupSocket = () => {
     if (globalThis.socketlib) {
         socketlibSocket = globalThis.socketlib.registerModule(MODULE_ID);
-        socketlibSocket.register("gmHandleSummon", handleSummonPlacingAndPost);
+        socketlibSocket.register("gmHandleSummon", handleSummonPlacementAndPost);
     }
     return !!globalThis.socketlib;
 };